+++ src/app/workspace/Workspace.js
@@ -1,8 +1,10 @@
 import React from "react";
 import ShallowComponent from "robe-react-commons/lib/components/ShallowComponent";
-import Content from "app/workspace/Content";
-import "react-notifications/lib/notifications.css";
-import NotificationContainer from "react-notifications/lib/NotificationContainer"; // eslint-disable-line import/no-unresolved
+import ToastContainer from "robe-react-ui/lib/toast/ToastContainer";
+import Header from "app/header/Header";
+import SideMenu from "robe-react-ui/lib/sidemenu/SideMenu";
+import Col from "react-bootstrap/lib/Col";
+import Card from "libs/card/Card";
 
 export default class Workspace extends ShallowComponent {
 
@@ -10,15 +12,87 @@ export default class Workspace extends ShallowComponent {
         router: React.PropTypes.object.isRequired
     };
 
+    constructor(props:Object) {
+        super(props);
+        this.state = {
+            toggled: false
+        };
+    }
+
+
     render():Object {
+        let toggled = this.state.toggled == false ? 0 : 290;
         return (
             <div>
-                <NotificationContainer />
-                <Content
-                    content={this.props.children}
-                    router={this.context.router}
-                    menu={this.props.route.menu}/>
+                <ToastContainer />
+                <Header
+                    matches={this.state.matches}
+                    toggled={this.state.toggled}
+                    onToggle={this.__changeMenu}/>
+                <Col
+                    id="sideMenu"
+                    style={{width:toggled}}
+                    className="side-menu">
+                    <Card style={{marginLeft:0}}>
+                        <SideMenu
+                            items={this.props.route.menu[0]}
+                            selectedItem={"Dashboard"}
+                            onChange={this.__handleChange}/>
+                    </Card>
+                </Col>
+                <Col
+                    id="content"
+                    className="content"
+                    style={{ height:window.innerHeight-80,marginLeft:toggled}}
+                    onClick={this.__closeMenu}>
+                    {this.props.children}
+                </Col>
             </div>
         );
+    }
+
+    __handleChange = (item:Object)=> {
+        this.context.router.push(item.module);
+    };
+
+    __closeMenu = ()=> {
+        if (this.state.matches == true) {
+            this.setState({
+                toggled: false
+            });
+        }
+    };
+    __changeMenu = ()=> {
+        if (this.state.matches == true) {
+            this.setState({
+                toggled: !this.state.toggled
+            });
+        }
+    };
+
+    __mediaQueryChanged = (mql)=> {
+        this.setState({
+            toggled: !mql.matches,
+            matches: mql.matches
+        });
+
+    };
+
+    componentWillMount() {
+        const mql = window.matchMedia("screen and (max-width: 768px)");
+        mql.addListener(this.__mediaQueryChanged);
+        this.setState({matches: mql.matches, toggled: !mql.matches});
+
+        this.context.router.listen(this.__closeMenu);
+
+        let initialSelection = window.location.pathname.slice(window.location.pathname.lastIndexOf("/") + 1);
+        if (initialSelection) {
+            this.context.router.push(initialSelection);
+        }
+    }
+
+    componentWillUnmount() {
+        this.state.mql.removeListener(this.__mediaQueryChanged);
+        this.context.router.listen(null);
     }
 }